requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.2
opencv-python==4.11.0.86
pillow==11.1.0
psycopg==3.2.4
python-dotenv==1.0.1
pyzbar==0.1.9
qrcode==8.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
